feat: add stats context provider
