+++ src/hooks/useStats.js
@@ -1,46 +1,1 @@
-import { useState, useCallback } from 'react';
-import { getMyStats } from '../services/stats';
-import { syncVouchers } from '../services/vouchers';
-import { applyStampAccrual } from '../utils/rewards';
-
-export function useStats() {
-  const initial = globalThis.preloaded?.stats || { loyaltyStamps: 0, freebiesLeft: 0, vouchers: [] };
-  const [stats, setStats] = useState(initial);
-
-  const refreshStats = useCallback(async (force = false) => {
-    if (!force && globalThis.preloaded?.stats) {
-      setStats(globalThis.preloaded.stats);
-      return globalThis.preloaded.stats;
-    }
-    try {
-      let s = await getMyStats();
-      const mismatch = s.freebiesLeft !== (Array.isArray(s.vouchers) ? s.vouchers.length : 0);
-      const outOfRange = s.loyaltyStamps < 0 || s.loyaltyStamps > 7;
-      if (mismatch || outOfRange) {
-        await syncVouchers();
-        s = await getMyStats();
-      }
-      if (s.loyaltyStamps < 0 || s.loyaltyStamps > 7) {
-        const { vouchersEarned, stampsRemainder } = applyStampAccrual(0, s.loyaltyStamps);
-        s.loyaltyStamps = stampsRemainder;
-        if (vouchersEarned > 0) {
-          s.freebiesLeft += vouchersEarned;
-          s.vouchers = Array.isArray(s.vouchers) ? s.vouchers : [];
-        }
-      }
-      setStats(s);
-      globalThis.freebiesLeft = s.freebiesLeft;
-      globalThis.loyaltyStamps = s.loyaltyStamps;
-      globalThis.preloaded = globalThis.preloaded || {};
-      globalThis.preloaded.stats = s;
-      console.log('loyalty stamps and free drinks have been received');
-      return s;
-    } catch {
-      const fallback = { loyaltyStamps: 0, freebiesLeft: 0, vouchers: [] };
-      setStats(fallback);
-      return fallback;
-    }
-  }, []);
-
-  return { stats, refreshStats };
-}
+export { StatsContext, StatsProvider } from '../context/StatsContext';
